move serialization test back
